alta y modificacion cursos

+++ public/js/index.js
@@ -1,19 +1,75 @@
+// data inicial
+var initialData = {
+	listadoDocentes : [],
+	listadoAulas : [],
+	listadoMaterias : []
+}
+
 $(document).ready(()=>{
 	var termina = {
+		infoCargada: false,
 		cursosCargados: false
 	};
 
 	$.blockUI({message: "Cargando.."})
 	cargarCursos(termina);
+	cargarInfo(termina);
 });
 
 function borrarCurso(curso){
-	$.ajax({
+	swal({
+	  title: 'Estás seguro?',
+	  text: "No vas a poder revertir esta acción!",
+	  type: 'warning',
+	  showCancelButton: true,
+	  confirmButtonColor: '#d33',
+	  cancelButtonColor: '#3085d6',
+	  confirmButtonText: 'Si, borralo!',
+	  cancelButtonText: 'Cancelar'
+	}).then((result) => {
+	  if (result.value) {
+	  	$.ajax({
 			url: '../admin/cursos/' + curso,
 			type: 'DELETE',
 			success: (data)=>{
-				console.log(data);
+				swal(
+			      'Borrado!',
+			      'El curso se ha eliminado correctamente.',
+			      'success'
+			    )
 				cargarCursos();
+			}
+		});
+	  }
+	})	
+}
+
+function cargarInfo(termina){
+	$.ajax({
+			url: '../admin/info',
+			type: 'GET',
+			success: (data)=>{
+				if (data){
+					if(data[0]){
+						initialData.listadoDocentes = data[0].rows;
+						for (var i = 0; i < data[0].rows.length; i++) {
+							var docente = data[0].rows[i];
+							var option = new Option(docente.nombres, docente.legajo, false, false);
+							$("#modalSelectDocentes").append(option);
+						}
+						$("#modalSelectDocentes").trigger('change');
+					}
+					if(data[1]){
+						initialData.listadoAulas = data[1].rows;
+					} 
+				}
+
+				if(termina){
+					termina.infoCargada = true;
+					if(termina.cursosCargados && termina.infoCargada){
+						$.unblockUI();
+					}
+				}
 			}
 		});
 }
@@ -23,7 +79,6 @@ function cargarCursos(termina){
 			url: '../admin/cursos',
 			type: 'GET',
 			success: (data)=>{
-				console.log(data)
 				if (data.length > 0){
 					$("#tablaCursos").empty();
 
@@ -81,34 +136,43 @@ function cargarCursos(termina){
 						tbody.append(tr);
 
 						var tdCodigo = document.createElement("td");
+						tdCodigo.setAttribute("data-info", "codigo");
 						tdCodigo.innerHTML = data[i].codigo;
 						tr.append(tdCodigo);
 
 						var tdNombre = document.createElement("td");
+						tdNombre.setAttribute("data-info", "nombre");
 						tdNombre.innerHTML = data[i].nombre;
 						tr.append(tdNombre);
 
 						var tdDocente = document.createElement("td");
+						tdDocente.setAttribute("data-info", "docente");
+						tdDocente.setAttribute("data-legajo", data[i].docente_a_cargo);
 						tdDocente.innerHTML = data[i].nombre_docente;
 						tr.append(tdDocente);
 
 						var tdSede = document.createElement("td");
+						tdSede.setAttribute("data-info", "sedes");
 						tdSede.innerHTML = data[i].sede;
 						tr.append(tdSede);
 
 						var tdAulas = document.createElement("td");
+						tdAulas.setAttribute("data-info", "aulas");
 						tdAulas.innerHTML = data[i].aulas;
 						tr.append(tdAulas);
 
 						var tdCupos = document.createElement("td");
+						tdCupos.setAttribute("data-info", "cupos");
 						tdCupos.innerHTML = data[i].cupos_disponibles;
 						tr.append(tdCupos);
 
 						var tdDias = document.createElement("td");
+						tdDias.setAttribute("data-info", "dias");
 						tdDias.innerHTML = data[i].dias;
 						tr.append(tdDias);
 
 						var tdHorarios = document.createElement("td");
+						tdHorarios.setAttribute("data-info", "horarios");
 						tdHorarios.innerHTML = data[i].horarios;
 						tr.append(tdHorarios);
 
@@ -120,6 +184,10 @@ function cargarCursos(termina){
 						var spanEditar = document.createElement("span");
 						spanEditar.className = "glyphicon glyphicon-edit";
 						spanEditar.setAttribute("aria-hidden", "true");
+						spanEditar.setAttribute("data-toggle", "modal");
+						spanEditar.setAttribute("data-target", "#modalCurso");
+						spanEditar.setAttribute("data-accion", "editar");
+						spanEditar.setAttribute("data-id", data[i].id_curso);
 						tdAcciones.append(divEditar)
 						divEditar.append(spanEditar);
 
@@ -140,7 +208,7 @@ function cargarCursos(termina){
 
 				if(termina){
 					termina.cursosCargados = true;
-					if(termina.cursosCargados){
+					if(termina.cursosCargados && termina.infoCargada){
 						$.unblockUI();
 					}
 				}
@@ -259,6 +327,211 @@ function clickDocentes() {
 	}
 }
 
+//manejo modal
+function cerrarModal(){
+	swal({
+	  title: 'Estás seguro?',
+	  type: 'warning',
+	  showCancelButton: true,
+	  confirmButtonColor: '#d33',
+	  cancelButtonColor: '#3085d6',
+	  confirmButtonText: 'Si',
+	  cancelButtonText: 'Cancelar'
+	}).then((result) => {
+	  if (result.value) {
+	  	$("#modalCurso").modal('hide');
+	  }
+	})	
+}
+
+function guardarCurso(){
+	var docente = $("#modalSelectDocentes").val();
+	var dias = $("#modalSelectDias").val();
+	var strDias = "", strAulas = "", strSedes = "", strHorarios = "";
+	for (var i = 0; i < dias.length; i++) {
+		strAulas = strAulas + $("#modalSelectAulas" + dias[i]).val() + ";";
+		strSedes = strSedes + $("#modalSelectSedes" + dias[i]).val() + ";";
+		strDias = strDias + dias[i] + ";";
+	}
+
+	strAulas = strAulas.slice(0, -1);
+	strSedes = strSedes.slice(0, -1);
+	strDias = strDias.slice(0, -1);
+
+	var idCurso = $("#modalInputId").val();
+	$.blockUI({message:"Guardando.."})
+	$.ajax({
+			url: '../admin/cursos/' + idCurso,
+			type: 'PUT',
+			data: {
+				dias: strDias,
+				aulas: strAulas,
+				sedes: strSedes
+			},
+			success: (data)=>{
+				$.unblockUI();
+				swal({
+				  type: 'success',
+				  title: 'Guardado!',
+				  text: 'Se han guardado los cambios'
+				});
+				$("#modalCurso").modal('hide');
+				cargarCursos();
+			},
+			error: function (xhr, ajaxOptions, thrownError) {
+		        $.unblockUI();
+		        swal({
+				  type: 'error',
+				  title: 'Oops...',
+				  text: 'Ha ocurrido un error al intentar guardar curso'
+				});
+		        console.log(xhr.responseText);
+		        console.log(thrownError);
+	      	}
+		});
+}
+
+function agregarCurso(){
+	var docente = $("#modalSelectDocentes").val();
+	var dias = $("#modalSelectDias").val();
+	var strDias = "", strAulas = "", strSedes = "", strHorarios = "";
+	for (var i = 0; i < dias.length; i++) {
+		strAulas = strAulas + $("#modalSelectAulas" + dias[i]).val() + ";";
+		strSedes = strSedes + $("#modalSelectSedes" + dias[i]).val() + ";";
+		strDias = strDias + dias[i] + ";";
+	}
+
+	strAulas = strAulas.slice(0, -1);
+	strSedes = strSedes.slice(0, -1);
+	strDias = strDias.slice(0, -1);
+
+	var idCurso = $("#modalInputId").val();
+	$.blockUI({message:"Guardando.."})
+	$.ajax({
+			url: '../admin/cursos/',
+			type: 'POST',
+			data: {
+				dias: strDias,
+				aulas: strAulas,
+				sedes: strSedes
+			},
+			success: (data)=>{
+				$.unblockUI();
+				swal({
+				  type: 'success',
+				  title: 'Guardado!',
+				  text: 'Se han guardado los cambios'
+				});
+				$("#modalCurso").modal('hide');
+				cargarCursos();
+			},
+			error: function (xhr, ajaxOptions, thrownError) {
+		        $.unblockUI();
+		        swal({
+				  type: 'error',
+				  title: 'Oops...',
+				  text: 'Ha ocurrido un error al intentar guardar curso'
+				});
+		        console.log(xhr.responseText);
+		        console.log(thrownError);
+	      	}
+		});
+}
+
+$('#modalCurso').on('show.bs.modal', function (event) {
+	$("#rowDia").html('');
+	$("#modalSelectMaterias").select2({ width: '100%' });
+	$("#modalSelectDocentes").select2({ width: '100%' });
+	$("#modalSelectDias").select2({ width: '100%' });
+	var button = $(event.relatedTarget); // Button that triggered the modal
+	var accion = button.data('accion'); // Extract info from data-* attributes
+	var modal = $(this);
+
+  	if (accion == "editar") {
+  		modal.find('.modal-title').text('Editar Curso');
+
+  		$("#btnModal").attr("onclick", "guardarCurso()");
+
+  		//obtengo toda la info del curso seleccionado
+  		var tr = $("tr[id='" + button.data('id') + "']");
+
+  		$("#modalInputId").val(button.data('id'));
+
+  		var docente = tr.find("td[data-info='docente']").attr("data-legajo");
+  		$("#modalSelectDocentes").val(docente).trigger("change");
+
+  		var dias = tr.find("td[data-info='dias']")[0].innerHTML.trim().split(";");
+  		$("#modalSelectDias").val(dias).trigger("change");
+
+  		var sedes = tr.find("td[data-info='sedes']")[0].innerHTML.trim().split(";");
+  		for (var i = 0; i < dias.length; i++) {
+  			$("#modalSelectSedes" + dias[i]).val(sedes[i]).trigger("change");
+  		}
+
+  		var aulas = tr.find("td[data-info='aulas']")[0].innerHTML.trim().split(";");
+  		for (var i = 0; i < dias.length; i++) {
+  			$("#modalSelectAulas" + dias[i]).val(aulas[i]).trigger("change");
+  		}
+  	} else{
+  		modal.find('.modal-title').text('Agregar Curso');
+
+  		$("#btnModal").attr("onclick", "agregarCurso()");
+  	}
+})
+
+$("#modalSelectDias").on('select2:unselect', (event) => {
+	$("#rowDia" + event.params.data.id).remove();
+});
+
+$("#modalSelectDias").on('change', (event) => {
+	//$("#rowDia").html('');
+	var dias = $("#modalSelectDias").select2('data');
+	for (var i = 0; i < dias.length; i++) {
+		if(!$("#rowDia" + dias[i].id).length){
+			$("#rowDia").append("\
+				<div class='row' id='rowDia" + dias[i].id + "'>\
+					<div class='col-lg-1'></div>\
+					<div class='col-lg-10'>\
+						<h4 align='center'>Dia " + dias[i].text + "</h4>\
+						<div class='row'>\
+							<div class='col-lg-12'>\
+							    <label>Sede:</label>\
+							    <select id='modalSelectSedes" + dias[i].id + "'>\
+							    	<option value='LH'>Las Heras</option>\
+							    	<option value='PC'>Paseo Colón</option>\
+							    	<option value='CU'>Ciudad Universitaria</option>\
+							    </select>\
+							</div>\
+						</div>\
+						\
+						<div class='row'>\
+							<div class='col-lg-12'>\
+							    <label>Horarios:</label>\
+							    <input type='text' id='modalSelectHorarios" + dias[i].id + "' style='width: 100%'>\
+							</div>\
+						</div>\
+						<div class='row'>\
+							<div class='col-lg-12'>\
+							    <label>Aulas:</label>\
+							    <select id='modalSelectAulas" + dias[i].id + "'>\
+							    	\
+							    </select>\
+							</div>\
+						</div>\
+					</div>\
+				</div>");
+			$("#modalSelectSedes" + dias[i].id).select2({ width: '100%' });
+			$("#modalSelectAulas" + dias[i].id).select2({ width: '100%' });
+			for (var j = 0; j < initialData.listadoAulas.length; j++) {
+				var aula = initialData.listadoAulas[j];
+				var option = new Option(aula.aula, aula.aula, false, false);
+				$("#modalSelectAulas" + dias[i].id).append(option);
+			}
+			$("#modalSelectAulas" + dias[i].id).trigger('change');
+		}
+	}
+});
+
 // maneja menu
 function cambiarPantalla(numero){
 	switch(numero){
@@ -331,3 +604,11 @@ $(".input-file-docentes").before(
 	}
 	
 );
+
+$('select').on('select2:select', function(e){
+    var elm = e.params.data.element;
+    $elm = jQuery(elm);
+    $t = jQuery(this);
+    $t.append($elm);
+    $t.trigger('change.select2');
+});
